+++ src/pages/dashboard/adminDashboard.tsx
@@ -1,0 +1,7 @@
+export function Dashboard() {
+  return (
+    <div className='flex flex-col justify-center items-center'>
+      <span className='text-3xl underline'>Tableau de bord administrateur</span>
+    </div>
+  );
+}