+++ src/pages/home.tsx
@@ -1,16 +1,15 @@
-import './_home.scss'
+import './_home.scss';
 
-import rocket from '../assets/images/header_rocket.webp'
-import rocketSmoke from '../assets/images/header_rocket_smoke.webp'
-import rocketWave from '../assets/images/header_wave.svg'
-import mouseScroll from '../assets/icons/mouse_scroll.svg'
+import rocket from '../assets/images/header_rocket.webp';
+import rocketSmoke from '../assets/images/header_rocket_smoke.webp';
+import rocketWave from '../assets/images/header_wave.svg';
+import mouseScroll from '../assets/icons/mouse_scroll.svg';
 
-import { ConsoleCommand } from '../components/consoleCommand'
-import ServiceSection from '../components/serviceSection'
-import PresentationSection from '../components/presentationSection'
-import PortfolioSection from '../components/portfolioSection'
-import ContactSection from '../components/contactSection'
-import TestSection from '../components/testSection'
+import { ConsoleCommand } from '../components/consoleCommand';
+import ServiceSection from '../components/serviceSection';
+import PresentationSection from '../components/presentationSection';
+import PortfolioSection from '../components/portfolioSection';
+import ContactSection from '../components/contactSection';
 
 export default function HomePage() {
   return (
@@ -30,9 +29,9 @@ export default function HomePage() {
         <img src={mouseScroll} alt='' className='mouse-scroll' />
       </section>
       {/* <ServiceSection />
-      <PresentationSection />
       <PortfolioSection /> */}
+      <PresentationSection />
       <ContactSection />
     </>
-  )
+  );
 }